remove conversions to CUDA API types from main library, update RMM to include @nvidia/cuda headers
