+++ modules/cuda/src/casting.hpp
@@ -19,8 +19,8 @@
 
 // #include <cuda_runtime_api.h>
 // #include <nv_node/utilities/args.hpp>
-// #include "cuda/utilities/napi_to_cpp.hpp"
-// #include "cuda/utilities/cpp_to_napi.hpp"
+// #include "node_cuda/utilities/napi_to_cpp.hpp"
+// #include "node_cuda/utilities/cpp_to_napi.hpp"
 
 // #include <type_traits>
