+++ modules/cuda/src/mem.cpp
@@ -13,9 +13,9 @@
 // limitations under the License.
 
 #include "buffer.hpp"
-#include "cuda/utilities/cpp_to_napi.hpp"
-#include "cuda/utilities/napi_to_cpp.hpp"
-#include "macros.hpp"
+#include "node_cuda/macros.hpp"
+#include "node_cuda/utilities/cpp_to_napi.hpp"
+#include "node_cuda/utilities/napi_to_cpp.hpp"
 #include "task.hpp"
 
 #include <cuda.h>
